Fixed issue with resetting zoom after switching currencies
I Determined that the cause of the bug was that trades weren't being cleared after switching currencies.  The new min/max prices were being correctly calculated and passed down into the orderbook component,
   but the `resetZoom()` functionality was ignoring them since there were existing trades.

To solve this, I manually reset the trade history every time the currency changes and then call the `resetZoom()` function (which was modified slightly to work even if there is no drawn button)
to reset the view to defaults for the new currency.